+++ bootstrap/cache/packages.php
@@ -1,0 +1,246 @@
+<?php return array (
+  'barryvdh/laravel-debugbar' => 
+  array (
+    'aliases' => 
+    array (
+      'Debugbar' => 'Fruitcake\\LaravelDebugbar\\Facades\\Debugbar',
+    ),
+    'providers' => 
+    array (
+      0 => 'Fruitcake\\LaravelDebugbar\\ServiceProvider',
+    ),
+  ),
+  'bkwld/croppa' => 
+  array (
+    'aliases' => 
+    array (
+      'Croppa' => 'Bkwld\\Croppa\\Facades\\Croppa',
+    ),
+    'providers' => 
+    array (
+      0 => 'Bkwld\\Croppa\\CroppaServiceProvider',
+    ),
+  ),
+  'laracasts/utilities' => 
+  array (
+    'aliases' => 
+    array (
+      'JavaScript' => 'Laracasts\\Utilities\\JavaScript\\JavaScriptFacade',
+    ),
+    'providers' => 
+    array (
+      0 => 'Laracasts\\Utilities\\JavaScript\\JavaScriptServiceProvider',
+    ),
+  ),
+  'laravel/boost' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Laravel\\Boost\\BoostServiceProvider',
+    ),
+  ),
+  'laravel/mcp' => 
+  array (
+    'aliases' => 
+    array (
+      'Mcp' => 'Laravel\\Mcp\\Facades\\Mcp',
+    ),
+    'providers' => 
+    array (
+      0 => 'Laravel\\Mcp\\Server\\McpServiceProvider',
+    ),
+  ),
+  'laravel/pail' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Laravel\\Pail\\PailServiceProvider',
+    ),
+  ),
+  'laravel/roster' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Laravel\\Roster\\RosterServiceProvider',
+    ),
+  ),
+  'laravel/sail' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Laravel\\Sail\\SailServiceProvider',
+    ),
+  ),
+  'laravel/tinker' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Laravel\\Tinker\\TinkerServiceProvider',
+    ),
+  ),
+  'nesbot/carbon' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Carbon\\Laravel\\ServiceProvider',
+    ),
+  ),
+  'nunomaduro/collision' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'NunoMaduro\\Collision\\Adapters\\Laravel\\CollisionServiceProvider',
+    ),
+  ),
+  'nunomaduro/termwind' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Termwind\\Laravel\\TermwindServiceProvider',
+    ),
+  ),
+  'spatie/eloquent-sortable' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\EloquentSortable\\EloquentSortableServiceProvider',
+    ),
+  ),
+  'spatie/laravel-feed' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\Feed\\FeedServiceProvider',
+    ),
+  ),
+  'spatie/laravel-ignition' => 
+  array (
+    'aliases' => 
+    array (
+      'Flare' => 'Spatie\\LaravelIgnition\\Facades\\Flare',
+    ),
+    'providers' => 
+    array (
+      0 => 'Spatie\\LaravelIgnition\\IgnitionServiceProvider',
+    ),
+  ),
+  'spatie/laravel-markdown-response' => 
+  array (
+    'aliases' => 
+    array (
+      'Markdown' => 'Spatie\\MarkdownResponse\\Facades\\Markdown',
+    ),
+    'providers' => 
+    array (
+      0 => 'Spatie\\MarkdownResponse\\MarkdownResponseServiceProvider',
+    ),
+  ),
+  'spatie/laravel-one-time-passwords' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\OneTimePasswords\\OneTimePasswordsServiceProvider',
+    ),
+  ),
+  'spatie/laravel-passkeys' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\LaravelPasskeys\\LaravelPasskeysServiceProvider',
+    ),
+  ),
+  'spatie/laravel-permission' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\Permission\\PermissionServiceProvider',
+    ),
+  ),
+  'spatie/laravel-query-builder' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Spatie\\QueryBuilder\\QueryBuilderServiceProvider',
+    ),
+  ),
+  'spatie/laravel-responsecache' => 
+  array (
+    'aliases' => 
+    array (
+      'ResponseCache' => 'Spatie\\ResponseCache\\Facades\\ResponseCache',
+    ),
+    'providers' => 
+    array (
+      0 => 'Spatie\\ResponseCache\\ResponseCacheServiceProvider',
+    ),
+  ),
+  'typicms/bootforms' => 
+  array (
+    'aliases' => 
+    array (
+      'BootForm' => 'TypiCMS\\BootForms\\Facades\\BootForm',
+    ),
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\BootForms\\BootFormsServiceProvider',
+    ),
+  ),
+  'typicms/core' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\Modules\\Core\\Providers\\ModuleServiceProvider',
+      1 => 'TypiCMS\\Modules\\Core\\Providers\\TranslationsServiceProvider',
+    ),
+  ),
+  'typicms/form' => 
+  array (
+    'aliases' => 
+    array (
+      'Form' => 'TypiCMS\\Form\\Facades\\Form',
+    ),
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\Form\\FormServiceProvider',
+    ),
+  ),
+  'typicms/laravel-translatable-bootforms' => 
+  array (
+    'aliases' => 
+    array (
+      'TranslatableBootForm' => 'TypiCMS\\LaravelTranslatableBootForms\\Facades\\TranslatableBootForm',
+    ),
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\LaravelTranslatableBootForms\\TranslatableBootFormsServiceProvider',
+    ),
+  ),
+  'typicms/sidebar' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\Modules\\Sidebar\\SidebarServiceProvider',
+    ),
+  ),
+  'typicms/translatable' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'TypiCMS\\Translatable\\TranslatableServiceProvider',
+    ),
+  ),
+  'typidesign/laravel-artisan-translations' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Typidesign\\Translations\\ArtisanTranslationsServiceProvider',
+    ),
+  ),
+  'ultrono/laravel-sitemap' => 
+  array (
+    'providers' => 
+    array (
+      0 => 'Ultrono\\Sitemap\\SitemapServiceProvider',
+    ),
+  ),
+);