+++ database/migrations/2025_08_18_152546_create_passkeys_table.php
@@ -5,15 +5,15 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Spatie\LaravelPasskeys\Support\Config;
 
-return new class extends Migration {
-
+return new class extends Migration
+{
     public function up(): void
     {
         $authenticatableClass = Config::getAuthenticatableModel();
 
-        $authenticatableTableName = (new $authenticatableClass)->getTable();
+        $authenticatableTableName = (new $authenticatableClass())->getTable();
 
-        Schema::create('passkeys', function (Blueprint $table) use ($authenticatableTableName,$authenticatableClass) {
+        Schema::create('passkeys', function (Blueprint $table) use ($authenticatableTableName, $authenticatableClass) {
             $table->id();
 
             $table