+++ lang/en/languages.php
@@ -1,10 +1,7 @@
 <?php
 
 return [
-<<<<<<< Updated upstream
     'ar' => 'Arabic',
-=======
->>>>>>> Stashed changes
     'fr' => 'French',
     'es' => 'Spanish',
     'nl' => 'Dutch',