+++ resources/js/admin.js
@@ -28,21 +28,10 @@ import { createApp } from 'vue';
  * i18n
  */
 import { createI18n } from 'vue-i18n';
-<<<<<<< Updated upstream
 import ar from '../../lang/ar.json';
 import en from '../../lang/en.json';
-import es from '../../lang/es.json';
-import fr from '../../lang/fr.json';
-import nl from '../../lang/nl.json';
 
-const messages = { ar, fr, en, es, nl };
-=======
-import en from '../../lang/en.json';
-import es from '../../lang/es.json';
-import fr from '../../lang/fr.json';
-
-const messages = { fr, en, es };
->>>>>>> Stashed changes
+const messages = { ar, en };
 const i18n = new createI18n({
     legacy: false,
     locale: window.TypiCMS.locale,