+++ vite.config.js
@@ -4,18 +4,26 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
     build: {
-        chunkSizeWarningLimit: 2000,
+        chunkSizeWarningLimit: 3000,
+    },
+    css: {
+        preprocessorOptions: {
+            scss: {
+                silenceDeprecations: ['import', 'global-builtin', 'color-functions', 'if-function'],
+            },
+        },
+        devSourcemap: true,
     },
     plugins: [
         laravel({
             input: ['resources/scss/public.scss', 'resources/scss/admin.scss', 'resources/js/public.js', 'resources/js/admin.js', 'resources/js/admin/theme-switcher.ts'],
             refresh: true,
+            assets: ['resources/images/**'],
         }),
         vue({
             template: {
-                transformAssetUrls: {
-                    // base: null,
-                    // includeAbsolute: false,
+                compilerOptions: {
+                    isCustomElement: (tag) => tag.startsWith('cropper-'),
                 },
             },
         }),